add DB connection, add logic to get book from DB

+++ src/main/java/com/atlas/library/bookmanagement/model/Book.java
@@ -18,7 +18,7 @@ public class Book {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Long bookId;
+    private int bookId;
 
     @Column(unique = true)
     private String isbn;
